+++ .gitignore
@@ -1,1 +1,1 @@
-__pycache__/*
+src/__pycache__/*
